refactor(flui-layer): Mythos Step 10 -- split compositor.rs into directory
Decompose the 967 LOC `compositor.rs` god module into a directory of
focused files, each with a single concern.

LOC delta:

- Delete: `crates/flui-layer/src/compositor.rs` (967 LOC -- 600 LOC of
  SceneBuilder + 80 LOC of SceneCompositor + 287 LOC of inline tests).
- Create: `crates/flui-layer/src/compositor/mod.rs` (15 LOC of re-exports).
- Create: `crates/flui-layer/src/compositor/builder.rs` (~557 LOC --
  SceneBuilder<'a> + all push_*/add_*/pop/build/finish methods).
- Create: `crates/flui-layer/src/compositor/retained.rs` (~96 LOC --
  SceneCompositor + CompositorStats; structurally separate from
  SceneBuilder).
- Create: `crates/flui-layer/tests/scene_builder.rs` (~285 LOC --
  integration tests for both SceneBuilder and SceneCompositor).
- Net: roughly equivalent total LOC; concerns separated.

Module layout after this step matches the verdict S6 module sketch:

  crates/flui-layer/src/compositor/
    mod.rs       -- re-exports
    builder.rs   -- SceneBuilder<'a>
    retained.rs  -- SceneCompositor + CompositorStats

The `crate::compositor` external paths (e.g. `flui_layer::SceneBuilder`,
`flui_layer::SceneCompositor`, `flui_layer::CompositorStats`) all
continue to resolve via the lib.rs re-export `pub use compositor::{...}`.

Integration tests:

- `crates/flui-layer/tests/scene_builder.rs` consumes the public crate
  API (`flui_layer::*`). The inline `mod tests {}` imports
  (`use super::*; use flui_painting::Canvas; use flui_types::...`)
  collapse to module-level imports. 17 tests pass (was 17 inline; same
  count, same scenarios).

Files touched:

- Delete: `crates/flui-layer/src/compositor.rs`.
- Create: `crates/flui-layer/src/compositor/{mod,builder,retained}.rs`.
- Create: `crates/flui-layer/tests/scene_builder.rs`.

Verification:

- `cargo check --workspace`: clean.
- `cargo test -p flui-layer --lib`: 229 passed, 0 failed (down from 246;
  17 SceneBuilder/SceneCompositor tests moved to integration).
- `cargo test -p flui-layer --tests`: 20 (layer_tree) + 17 (scene_builder)
  + 8 (damage_tracking pre-existing) = 45 integration tests pass.

External callers: all flui_rendering / flui_engine / flui_app / flui_hot_reload
re-exports through `flui_layer::*` resolve identically. No caller-side
changes required.

Plan: docs/plans/2026-05-20-002-feat-flui-layer-mythos-redesign-plan.md U10.
Verdict: docs/designs/2026-05-20-mythos-flui-layer-redesign.md S6.
